Trainer Traffic simulator.html
A project dedicated to nursery/primary school teachers to aid them in teaching their pupil in a more conducive and enticing manner.
during may days in the primary school, i was left to imagine the traffic colors as they were mentioned, which was boring and would make us feel sleepy. but i believe with projects like this, such lessons will fun and easily understood.